Review - pep8 fixes and improvements
I removed the unnecessary if check in config as all files
in the images directory were either jpeg or jpg files so
the check was redundant.

I moved the loading of images and initialization of the font
to the game constructor rather than when run is called.
it makes sense to load the images and initialize the font before
run is called.

Arrange imports according to pep8, and other pep8 fixes.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>